CrosEC: Force the function that updates the EC to use chunks of the right size.
The function had been writing data in chunks which were an appropriate size,
but the final bit would be whatever was left over. If the EC required that the
blobs written to the EC were multiples of the block size (as it does on Pit),
then that command would fail.

BUG=chrome-os-partner:19420
TEST=Built and booted on Pit and saw that the final write command to update
the EC image no longer failed with this change.
BRANCH=None

Change-Id: I0b4844c56b1070ab3e339eecfccb4541c4a546e1
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/62895
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>