sound: Add PCI device IDs for LynxPoint HDA
Add device IDs for LynxPoint-H and LynxPoint-LP for
the HDA sound driver.

BUG=chrome-os-partner:21216
BRANCH=falco
TEST=emerge-falco depthcharge

Note that this by itself it not enough to make it beep,
there still needs to be a beep verb table and possibly
some additional tweaks to depthcharge driver since it
expects some link-specific verb entries.

However with this change it no longer complains that it
cannot find the audio controller:
Audio: Controller not found!

Change-Id: Ifb42f9fa1f721de124634b7e99ef0ff1d4ee0057
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62900
Reviewed-by: Aaron Durbin <adurbin@chromium.org>