flash: spi: Restructure the drivers.
Separate the Exynos SPI driver from the flash drivers. The ICH drivers are
still flash drivers since they can't really be used for anything else, and
there's now a flash driver which calls into an underlying SPI driver for the
actual bus transactions.

The ICH flash, memory mapped flash, SPI flash, and exynos SPI drivers are now
all organized as structures of function pointers who's parameters are set when
structure is set up.

Also, the interface to SPI drivers is more in line with what's provided by the
hardware, so the code in the exynos SPI driver needed to be redistributed. In
the process the transfer function can now send, receive, or both, and it
decides whether to use 4 byte packets or not based on whether the total size
divides evenly by 4.

To keep the SPI bus running with 4 byte packets when loading from flash, the
SPI flash wrapper rounds the region being requested outward (start down, end
up) to ensure that the size is a multiple of 4 and you're getting at least the
data you're asking for. Ideally the driver would keep track of what it's
already loaded and not load it again, but that's not implemented as of now.

BUG=chrome-os-partner:19420
TEST=Built for all platforms except bolt. Booted on snow with the exynos SPI
driver, on Link with the memory mapped flash driver and also the ICH SPI flash
driver, booted RW on Link, and booted on Lumpy.
BRANCH=None

Original-Change-Id: Ib0348bd468896a569288d3ca2dfca534569195b9
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/62537
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
(cherry picked from commit d5e9febe56ceb0faa5b2b89052587544ea2d6f69)

Change-Id: I8fb4707dfc3843566616a88464281fea5f2387c1
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/63006