CrosEC: Add generic support for EC protocol version 3.
This adds the infrastructure to support protocol version 3, but doesn't
actually implement the required callback in any of the bus drivers. The
set_max_size functions turned out not to be useful since they'd adjust the
size of buffers internal to the bus drivers but the main part of protocol 3
is supported in generic code. This change makes the bus buffers allocated on
demand.

BUG=chrome-os-partner:19420
TEST=Built and booted into RW depthcharge on pit.
BRANCH=None

Change-Id: I4ec077d09ca4de6ce7312476ed52e933dd02e8ab
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/63194
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>