I2C TPM: If the TPM isn't open, open it.
When moving from the RO to the RW firmware, the I2C TPM driver doesn't
"reopen" the device and so the function pointers internal to the driver don't
get set up. When the firmware then tries to send a message to the TPM it
fails. This change adds some code to make sure that the TPM is open and also
adds some asserts to ensure we won't call into any NULL function pointers.

BUG=chrome-os-partner:19420
TEST=Built and booted an RW image on pit and made it into the kernel.
BRANCH=None

Change-Id: Ibd0b241fc959594bf8ddf961eebf9313de31a847
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/63195
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>