tpm: don't touch undefined register addresses
Certain TPMs don't behave well when any addresses not defined
in the PC client spec are accessed. There are two issues in the
current driver: access to the status and access register. The
access register is only 1 byte while the status register is 3 bytes.

The access register is handled easily by just using tpm_read_byte()
and tpm_write_byte(). Fixing the access to the status register is
more difficult because the burst count was previously pulled out
of the 32-bit value. Since the status register cannot be read be
read as a 32-bit value it is accessed out of line.

Lastly, tis_wait_reg() was changed to return an int with 0
being the success value.

BUG=chrome-os-partner:20564
BRANCH=None
TEST=Built and booted a bolt board which has a not-as-forgiving TPM.

Change-Id: I2d897244486cfe8b6dba4c789b03792cf0bf3eac
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/63250
Reviewed-by: Gabe Black <gabeblack@chromium.org>