sound: Turn the sound drivers into structures with function pointers.
Also introduce a wrapper class which represents the path audio takes through
the system called a SoundRoute. When you play a sound through the SoundRoute,
it first makes sure all of its components are enabled which would typically
include the codec, and then it calls through to the underlying sound source.

BUG=chrome-os-partner:19420
TEST=Built and booted on link and snow and verified that the developer mode
beep still works and the proceeded to the kernel. Built and booted on Falco
and saw consistent messages about not finding a beep node. Built for all other
platforms other than bolt. Built and booted on pit with no audio set up and
saw that it could still get through to the kernel with appropriate error
messages from the firmware.
BRANCH=none

Change-Id: Ifbe7909cfb632c70e6a03c2ab47f0444fdef22a9
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/63545
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>