exynos5250: gpio: Rename the GPIO driver from s5p to exynos5250.
The exynos5250 gpio driver had been called s5p. In addition to the fact that
I don't know what s5p is supposed to mean, it also implied that the driver was
more generic than it really was, considering there needed to be a separate
driver for the exynos5420.

BUG=chrome-os-partner:19420
TEST=Built and booted on snow.
BRANCH=None

Change-Id: I332d538b29b95941748ea5e5a49f8a10150c08e5
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/63557
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>