PS2: Disable keyboard before booting
Disable the keyboard controller and clear the keyboard buffer before
booting to the OS so that the linux driver doesn't get confused if
there is half a keystroke or something in the keyboard buffer.  This
was occasionally causing the keyboard to fail on boot.

This code could be removed at a later point if the EC code is changed
to clear the keyboard buffer on a kbc_reset.

BUG=chrome-os-partner:20758
TEST=Verified Code flow in normal boot/S3 resume with print statements.
     Verified Keyboard was correctly disabled and flushed by booting
     to recovery mode screen while pressing keys on the integrated
     keyboard.
BRANCH=none

Change-Id: I6bc17af963c64988beabd8dfb2d8fa16baca0613
Signed-off-by: Martin Roth <martin.roth@se-eng.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/63634
Reviewed-by: Gabe Black <gabeblack@chromium.org>