keyboard: When falsely detecting an escape sequence, still return the escape.
When the vboot keyboard callback looks at the keys coming in from getchar, it
assumes that escape is the start of an escape sequence. If escape isn't
followed by [, then the callback will give up on the escape sequence and throw
away the escape instead of returning it.

This code in general needs some work. If you type escape and the input driver
is blocking, the machine will hang waiting for the next keystroke to determine
if you're sending it an escape sequence. The keyboard drivers fortunately seem
to be non-blocking, but the serial driver is not. Also, if the character after
the escape is not a [, it will be dropped on the floor.

What we would do in an ideal world is call havechar for a while (maybe 10ms)
waiting for a character. If none shows up, give up and return the escape. If
one does, we should peek at it without actually pulling it from the input
device (we need a new hook for peeking) and, if it's not a [, leave it in the
queue for subsequent processing.

BUG=chrome-os-partner:21340
TEST=Built and booted on falco and saw that escape will now escape from the
transition to dev mode screen.
BRANCH=Haswell boards

Original-Change-Id: I8b50beb60d3837642f7c0953541f3d563271a15d
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/63516
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>

(cherry picked from commit 1d17624d9fb74782b94e6433f28854843821d7fa)
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>

Change-Id: I21f9e9572992b8089ac56b5330dd6aec927ac3fd
Reviewed-on: https://gerrit.chromium.org/gerrit/63927
Commit-Queue: Duncan Laurie <dlaurie@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Tested-by: Duncan Laurie <dlaurie@chromium.org>