exynos: video: Disable the FIMD channels as well as the outputs.
When disabling the FIMD so that it doesn't DMA while the sysmmu is being set
up, we were just disabling its output. That seems to have been sufficient for
the 5250 since snow boots with that fix. Due to a combination of other problems
it seemed that that worked for the 5420 as well, but in reality it didn't. This
change disables the channels feeding the FIMD explicitly as well as disabling
it's outputs.

BUG=chrome-os-partner:19420
TEST=Built for pit. Verified that this change gets rid of a panic happening
when the DRM driver is set up by the kernel.
BRANCH=None

Change-Id: I8f985e64185a56cf8492ad3ec5d2150b323ee7ae
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/64430
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>