EC: Fix a theoretical security problem in the i2c and spi send_command funcs.
The out_bytes and in_bytes variables in the i2c and spi send_command were
signed, but they were being assigned the value in a uint32_t with a constant
added in. If the value passed in is very large, the constant would make it
wrap to negative. A size check later in the function would pass since the
negative value would be lower than the max size, but would be very large if
treated as an unsigned value somewhere else.

BUG=chromium:270913
TEST=Built and booted on pit (which uses spi) using both version 2 and 3 of
the EC protocol. Verified that the EC id was read correctly. Built and booted
on snow (which uses i2c) and verified that the EC id was read correctly.
BRANCH=None

Change-Id: I44a667b4ac117179eb6fa93e38e2fa67a4e1a016
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/65503
Reviewed-by: Kees Cook <keescook@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>