I2C: TPM: Optimize delays in the I2C TPM driver.
If we check for completion every ms instead of every 5, we'll see the
completion sooner on average and spend less time hanging around for no reason.
Also, there was a hardcoded 2ms delay which was in the error case in one
and could have been intended for the error case in the other. The TPM seems
to work fine without those delays.

BUG=chrome-os-partner:19420
TEST=Built and booted several times on pit.
BRANCh=None

Change-Id: I784db2ffcbcbd658a9bb9fd3258e73a26cf9b3a7
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/66309
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Ronald G. Minnich <rminnich@chromium.org>
Reviewed-by: Ronald G. Minnich <rminnich@chromium.org>
Tested-by: Ronald G. Minnich <rminnich@chromium.org>