CHERRY-PICK: cleanup: Add and use new CleanupOnLegacy type for legacy boot
This adds a new CleanOnLegacy type that will be used when doing
a legacy boot rather than CleanupOnHandoff.  This allows the legacy
boot path to perform some cleanup actions but not others.

The existing arch_final_cleanup() for x86 is refactored to pull
out the work it was doing into cleanup handlers which are declared
as both CleanupOnLegacy and CleanupOnHandoff.  The PS/2 cleanup
is also modified to run in both cases.

There will be a new lynxpoint-specific cleanup function added in
the next commit which will be only used in the non-legacy path.

BUG=chrome-os-partner:22085
BRANCH=falco,peppy
TEST=manual: build and boot depthcharge on falco and see that the
current behavior is unchanged by this commit.  The MTRR cleanup,
the PS/2 keyboard cleanup, and the coreboot finalize step are still
executed properly in both legacy and non-legacy boot paths.

Original-Change-Id: I41e80aebec5aea94c972c7020e82362af500f844
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/66545
Reviewed-by: Gabe Black <gabeblack@chromium.org>
(cherry picked from commit 760b1ddbdc524031eb9769a9d35448361ba08f78)

Change-Id: I2ae516f278359becb20b643d774d45effd92d355
Signed-off-by: Marc Jones <marc.jones@se-eng.com>
Reviewed-on: https://chromium-review.googlesource.com/171215
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>