cleanup: Add a driver to route USB ports to XHCI
This driver will issue an SMI callback to coreboot that will
route all USB ports to the XHCI controller and disable the EHCI
controller to save power.  It is only used in the non-legacy
CleanupOnHandoff boot path.

In coreboot this is currently done as part of the finalize step
but it will be moved to the new callback so it can be limited to
the non-legacy path.

This driver is enabled on all haswell platforms.

BUG=chrome-os-partner:22085
BRANCH=falco,peppy
TEST=build and boot depthcharge with updated coreboot.  The USB
ports are routed to XHCI and the EHCI controller is disabled in
the normal/developer/recovery mode boot paths.  In the legacy
mode boot path the USB ports are left on the EHCI controller so
they can be used by SeaBIOS.

Original-Change-Id: I84c7357ac87afca602b9f25cd3aa3dc8e45e0b83
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/66546
Reviewed-by: Gabe Black <gabeblack@chromium.org>
(cherry picked from commit c5c105cca10735a98302d63f37af259b6d9125fc)

Change-Id: I6971b81d4b823833f144b2f373da97a53ddb74c9
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/66646