storage: Simplify the storage driver interface.
Instead of having an init callback and a refresh callback, there is now an
update call back which is called if need_update is set to 1. Controllers that
have fixed media attached to them will set need_update when their drivers are
initially set up and clear it after the first update. Controllers that might
have removable media attached will leave need_update turned on. Both will call
their own init function when necessary, probably the first time update is
called.

Also, to avoid initializing controllers which aren't attached to the types of
media we're concerned with, the single block_dev_controller list is split into
two lists, one for controllers that have fixed media and one for those that
could have removable media. The lists are not mutually exclusive, and the
same mechanism which determines when to initialize the controller on a single
list will prevent a controller for being initialized for each list it occurs
on.

BUG=None
TEST=Built for all boards. Built and booted from emmc and sd card on snow and
pit. Booted from ahci and usb on link. Booted from USB on kirby.
BRANCH=None

Change-Id: I134c207c61fa553a759bd54dcacbf5d93efb12e4
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/168044
Reviewed-by: ron minnich <rminnich@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>