i2c tpm: Remove mostly useless delay code/tables.
I assume from the code in the TPM driver that the TPM spec defines different
types of delays and timeouts which each have a particular duration, and that
the TPM can tell you how long each type is if you ask it. There was a large
table, some members of a data structure, and a function or two which managed
the timeouts and figured their value for different operations.

The timeout values for the various "ordinals" were never set in the vendor
specific data structure, however, and always defaulted to 2 minutes. Similarly
the timeouts a, b, c, and d were never overridden from their defaults. This
change gets rid of all the timeout management code and makes the "ordinal"
timeout 2 minutes and the a, b, c, and d timeouts 2 seconds, the larger of the
two default values.

Also, all of the affected loops with timeouts were converted away from the
style which loops a fixed number of times with a delay in the loop into the
more accurate style which waits until the timer says the timeout has passed.

BUG=None
TEST=Built and booted on kirby
BRANCH=None

Change-Id: Ifa46f060f76494b176ec686a7afdcc55078473fc
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/168363
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>