TPM: Stop requesting/releasing the TPM locality.
The locality is requested when the TPM is initialized and released when it's
cleaned up. There's no reason to set it to the same thing again and restore it
back to the same value before and after every transaction. Overall, this saves
something on the order of 10 to 20 ms when booting on pit which is small
enough to be difficult to measure reliably.

BUG=None
TEST=Built and booted on pit.
BRANCH=None

Change-Id: I2241337739995f83a31fee9bdaec94bc934181f0
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/168400
Reviewed-by: ron minnich <rminnich@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>