tpm: Move the i2c bus setting function out of the main tpm header.
The i2c bus setting function is only implemented or useful when using the i2c
tpm driver. This change moves it out of the main header and into one which is
included only if the i2c driver is going to be used.

BUG=None
TEST=Built for all supported boards.
BRANCH=None

Change-Id: I89f63be6aef1a7722ba775f560928e3691da1d48
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/168705
Reviewed-by: David Hendrix <dhendrix@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>