container_of: Split the container_of macro into its own header.
It used to be in the header with the linked list implementation. The
container_of macro was often needed in places that had nothing to do with
linked list, for instance in drivers. Rather than force people to include
list.h which is counterintuitive and brings in unneeded code, this change
splits container_of out into its own header file.

BUG=None
TEST=Built for all boards.
BRANCH=None

Change-Id: Idc4a77bd5152ad9ad6fe192cb1496f0c1dc8fe7b
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/168707
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: David Hendrix <dhendrix@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>