usb: Move USB host controller initialization to board.c
This change adapts depthcharge to the deprecation of the USB_MEMORY
mechanism in libpayload. It gains a new mechanism to define USB host
controllers through their type and base address with new_usb_hc(), which
will enqueue this information in a global list and use it with
libpayload's usb_add_mmio_hc() when USB is actually needed. Callers need
to make sure that the libpayload configuration matches depthcharge's
code such that each controller is only described in one of them.

CQ-DEPEND=CL:169454
BUG=chrome-os-partner:21969
TEST=Manual

Change-Id: Ibc51b17a2f84b20e4e352197614bdb78c0185dcd
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/169541
Reviewed-by: Gabe Black <gabeblack@chromium.org>