TPM: Turn the TPM drivers into structures of function pointers.
Narrow the TPM interface to just a transmit function. When the transmit
function is called for the first time the TPM is initialized. When that
happens, a cleanup function is installed which closes the TPM.

Rework and simplify the file structure of the I2C TPM driver. It's now split
into some generic bits and chip specific bits without also having a wrapper
layer between it and the real TPM interface.

BUG=None
TEST=Built for all boards. Booted on link, pit, and snow. Measured boot
performance on pit and verified it didn't change.
BRANCH=None

Change-Id: If1e40e706345ecd08fe4acbfd40d0c2e49306d14
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/169057
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>