fit: Make depthcharge consider more than just the first compat string.
When choosing a kernel device tree to boot, depthcharge was considering only
the first segment of the compatibility string. A recent change in the snow
device trees moved the string we wanted to match against to the second
position, breaking booting on that platform. An immediate solution would be to
update to whatever the first string is now, but that means things are less
flexible for the kernel going forward and would be best to avoid. Instead, we
now look at all segments of the string and pick the one that matches earliest.

Also the printing that shows what the compat strings were, which config was
picked and why are updated to make it clear what's going on.

BUG=None
TEST=Built and booted into a recent snow image. Booted on pit and verified
that there was no performance regression.
BRANCH=None

Change-Id: Ib2db40ad0146fe9e4b9b34ba04664c648e80da72
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/169457
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>