usb: Port Exynos5420 LoS bias hack ("PHY tuning") to depthcharge
This patch implements a software workaround for an Exynos5420 silicon
bug that is necessary to talk to SuperSpeed USB devices. Unfortunately,
there is really no where in the firmware stack this could fit without
being an ugly hack, so I'm just trying to go for least ugly here. See
comment in src/drivers/bus/usb/exynos.c for details.

BUG=chrome-os-partner:21969
TEST=Make sure SuperSpeed devices can be enumerated as such on Pit/Kirby

Change-Id: I5bbaf9314675e36eff2f42a0120899b11631bbef
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/173323
Reviewed-by: Gabe Black <gabeblack@chromium.org>