sysinfo: Stop automatically installing the sysinfo flags.
A collection of flags are passed into depthcharge from coreboot using the
coreboot tables and end up in the sysinfo structure. It's reasonable to use
the static value in some of those flags, but for some other values we need to
know their current and changing value. To support that and to avoid having to
install all the sysinfo GPIOs manually, a new function has been added to allow
replacing flags which were already installed. Since there's no guaranteed
ordering between init functions, the sysinfo flags need to be installed at the
same time as the other flags.

BUG=chrome-os-partner:22764
TEST=Built for all supported boards, booted on kirby.
BRANCH=None

Change-Id: Iec8340b65df5c63e446a51354c631dc9e0de487e
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/170000
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>