gpio: Add generic wrapper types which logically operate on regular gpios.
It might be necessary to read a gpio and normalize its polarity for use in
a generic bit of code, for instance code which checks whether or not the
power button is being pressed. It might also be necessary to combine two
values into one, for instance to combine the stored value of the power button
in the coreboot tables with the live value read directly from the GPIO.

To accomodate these use cases, several very simple gpio types have been added
which wrap other gpios and operate on their values. There's one which
logically inverts the value, one that logically ands the value of two gpios,
and one that logically ors them.

BUG=chrome-os-partner:22764
TEST=Built and booted on kirby, pit and snow. Booted into recovery mode with
the power switch flag hooked up through the inverting gpio wrapper and saw
that on all systems they shut down when the power button was pressed.
BRANCH=None

Change-Id: I97054fd8695408702d0916c781e7bf25a63660ca
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/170001
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>