ARM: Hook up the live power and lid switch values on each of the ARM boards.
On all systems we were relying on the value of the power and lid switch GPIOs
which were stored in the coreboot tables when vboot asked whether the user was
requesting to shut down the system. On x86 that worked because mechanisms set
up by coreboot would intervene when the power button was pressed or the lid
was closed which would shut down the system regardless of what we told vboot.
On ARM, however, what we returned mattered, and since what we were returning
was based on a cached value stored in a table, the user could press the power
button and/or close the lid all the wanted with no affect.

This change rearranges the way the flags are configured on each of the ARM
boards so that the live gpio values are returned instead of the cached sysinfo
value. The x86 boards are left as is at least for now since they would
probably never return true anyway since coreboot would have shut down the
system as soon as the power button or lid switch were detected.

BUG=chrome-os-partner:22764
TEST=Built and booted on pit, kirby, and snow and verified that the power
button or lid switch would make them to turn off while they waited at the
recovery screen.
BRANCH=None

Change-Id: I5dcd566a6a7453ef44281253dc2ce0be6dc92b05
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/170002
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>