ARM: fmap: Align the GBB on a 4K boundary.
A comment in the fmap said that the GBB should be aligned on a 4K boundary,
but the actual layout of the image had it at an offset. This caused some GBB
related utilities like gbb_set_flags.sh (used during FAFT tests) to misbehave
and wipe out other parts of the image, specifically the read only firmware id.

This change fixes things by switching the RO firmware id and the GBB while
keeping everything the same size and all the other sections in place. The
GBB's size isn't a multiple of 4K which is assumed to be ok.

BUG=None
TEST=Built and booted on pit with this change. Ran
/usr/share/vboot/bin/set_gbb_flags.sh and verified that the RO firmware ID was
still intact by rebooting and running crossystem. Before this change the RO
firmware ID would be replaced with all zeroes.
BRANCH=None

Change-Id: Icb6db326b54fce167ef83ae12e127669c9af1305
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/170003
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>