Kconfig: Fix the bayleybay board Kconfig include.
The bayleybay board was added recently along with an empty kconfig file (as
all boards have as of today) but didn't include that file in the kconfig tree.
That would be surprising if someone went to add an option for bayleybay and it
didn't show up, and was probably an oversight.

BUG=None
TEST=Built for bayleybay.
BRANCH=None

Change-Id: If7a093254d2dd358ced3b23d5d40d4584d08d703
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/170743
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>