netboot: Fix ASIX driver on AX88772
This patch adds some minor fixes and cleanups to the ASIX driver that
are needed to make it work with my AX88772-based Linksys dongle. It
removes some superfluous delays and extra resets in the init path, and
fixes the autonegotiation function to actually check for autonegotiation
support (not just that it's already active). It also adds a new
polling-based wait during init, since it appears that the PHY needs some
time after reset to become fully functional again, and will return only
zeroes for any MII reads before that.

BUG=None
TEST=Manual

Change-Id: I04ae74c64598ed96cec083d17abc9e09d10eb7f1
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/170795
Reviewed-by: Gabe Black <gabeblack@chromium.org>