netboot: Add ASIX network driver to boards where it had been forgotten
Only affects netboot images anyway, so there's no point in leaving out
the only network driver we have. I think it's pointless to have this
Kconfig at all only so that you can forget adding it to a new board and
confuse people, but I have to write it this way to make the upstream
maintainer happy.

BUG=None
TEST=Manual

Change-Id: Iefe7856b3b4ed41a4c26a34f7edadf5609963a54
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/170794
Reviewed-by: Gabe Black <gabeblack@chromium.org>