Trust EC software if there is none
Right now depthcharge will hang if trying to go into developer mode on a
system with no EC but a virtual developer mode switch.
This is because vboot asks whether it we trust the EC (which can not be
proven to exist). If there is no EC we might as well believe that
there's no malicious software running on it.

BUG=none
BRANCH=none
TEST=Go to dev mode on Beltino, and see that it works instead of hanging
     depthcharge.

Change-Id: I5910217e314e9408d4bd13e78e09c51b4225d651
Signed-off-by: Stefan Reinauer <reinauer@google.com>
Reviewed-on: https://chromium-review.googlesource.com/173608
Reviewed-by: Gabe Black <gabeblack@chromium.org>