tegra: dma: Add a driver for the tegra APB DMA channels.
These channels are managed by an object which handles the global enable and
allocating and deallocating the individual channels as other controllers need
them. Currently the only controller which would use them is the SPI
controller, but the logic to manage them as a pool is simple and it avoids
statically allocating channels to each driver.

BUG=None
TEST=With this and other changes, performed SPI reads and writes to the flash
using DMA. The write portion was sending the command and read command to the
chip.
BRANCH=None

Change-Id: Idb5b04bee6ced76eda2c44363005b0734e6f8a9b
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/173628
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>