tegra: spi: Add a SPI driver for tegra.
This driver is based on the one in coreboot and uses the APB DMA engine. It's
been reworked to fit into the depthcharge driver model, to remove #defines,
and to otherwise fit in with the code already in depthcharge.

The coreboot driver was still being developed when it was used as a basis for
this one, and the ability to use PIO to finish up oddly sized or small
transactions may be implemented differently. This change does its DMA
transactions using 4 byte packets. It does its PIO transactions using 1 byte
packets which allow it to handle stray bytes at the end of the transaction.

BUG=None
TEST=With this and other changes, built and booted into depthcharge on nyan.
Saw that several portions of the flash were read successfully including the
FMAP and portions of the GBB.
BRANCH=None

Change-Id: I162ad859643593883658dd4d4715ddbfaebfc376
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/173629
Reviewed-by: Ronald Minnich <rminnich@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>