nyan: Hook up the SPI flash.
This change instantiates data structures representing the pool of APB DMA
channels, the SPI controller attached to the flash, and the flash itself on
nyan, and installs the SPI flash instance as the target for flash operations.

BUG=None
TEST=Built and booted into depthcharge on nyan. Verified that the FMAP was
read correctly, or at least its signature was verified successfully and
various sections were found within it, and also that the GBB signature (and
some sections of it) were read successfully.
BRANCH=None

Change-Id: I8c9be4f71d43abc69318ce8d14fa796a776e1230
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/173711
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>