AS3722: Add a driver for the AS3722 PMIC which supports power ops.
I don't have documentation for the bitfields in the AS3722 registers, but the
force_reset and power_off bits in the ResetControl are very suggestively
named. The kernel driver uses the power_off bit to implement a power off
function which seems to justify using them this way. Emperically, these bits
have done what it sounds like they do.

BUG=None
TEST=With this and other changes, built and booted into depthcharge and saw
that it rebooted when it wanted to. Replaced the cold_reset callback with the
power_off one, and saw that when it wanted to reboot it powered off. Messages
on the EC console suggested that it noticed that the AP powered things down.
Put a loop with a printf after the write to the PMIC register. The loop
printed a few iterations and then stopped halfway through a message,
suggesting sudden power loss.
BRANCH=None

Change-Id: If538df249de2140c564704e082c440b592820df7
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/173831
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>