nyan: tegra: SPI: Add flow control, fix up constants, and fix waiting.
The driver had been waiting for the "go" bit to turn off to indicate that the
transfer was finished. This actually just means that the transfer has started,
and we need to wait for the number of packets sent to be equal to the number
we requested.

The mask constants used by the SPI and DMA driver were not shifted into
position. By pre-shifting them, they're a bit more concise to use. Also, the
maximum size for DMA transfers was one too high. The current value worked in
the block size register because it's set to n - 1, but when checking if the
transfer is done, the block count in the transfer status register is just n
and has to fit in 16 bits.

Finally, the driver didn't work correctly without flow control enabled in the
DMA controller which makes sense, considering there wouldn't otherwise be any
way for the two sides to coordinate.

BUG=None
TEST=With EC communication disabled in coreboot and the frequency adjusted
correctly, depthcharge was able to talk to the EC successfully.
BRANCH=None

Change-Id: I332d5f1db24e1550bfbc1ff2426739307a9d725a
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/174020
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>