nyan: Re-enable EC related options in the nyan config.
Now that communication with the EC works in depthcharge, these otions can be
re-enabled.

BUG=None
TEST=Built and booted into depthcharge on nyan. Saw that the version string
was read successfully. Also put in a test loop in main which reads key presses
and echoes them to the console. Pressed all the keys on the keyboard and
verified that the right character was printed.
BRANCH=None

Change-Id: Ieb0e1e0f21a751f4cb100512a28553bd5f384cef
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://chromium-review.googlesource.com/174021
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Hung-Te Lin <hungte@chromium.org>