HACK: Increase storage device spinup timeout
After power wash (or dev mode transition) completes, a certain SSD is
known to not respond to commands for ~13s. This can cause a timeout in
firmware, which leads to entering recovery mode. Work around this issue
by increasing the timeout to 20s, from 10s.

This change should be reverted once we properly resolve this issue with
a SSD firmware update (preferred) or possibly a change to the chromium
wiping script (clobber-state).

BUG=chrome-os-partner:23267
TEST=Manual. Run power wash on Peppy w/ 32 GB Kingston SSD. Confirm that
device does not reboot into recovery mode.

Change-Id: I9ddcc326778e50c97294df3aae5110fce5e658d9
Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/174889
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>